Possible covariance matrix fix
After extensive testing, it would appear the problem with the covariance matrix becoming non positive-semi-definite may simply be due to lack of numerical precision.  In order to counter this, I've added a P = (P + P^T)/2 to every measurement update, which seems to have fixed the problem.  However further investigation is warrented.