Working on square root filtering
Have a rough implementation of the time update and measurement update square root implementations.

But, we also need a way to update the sum of covariances (from their decompositions).  I think the way to do this is through the gram-schmidt (or other) orthogonalization trick, i.e. what Simon mentions in his chapter on square root filtering.   On paper it seems that this method might work, but I need to verify.