More work on square root filtering
In particular, implementing the new covararianceContainer class in the substates.

Also, changed the way the spread of means term is computed in the JPDAF update if cholesky decomposition is used.

So far, comparisons between the two methods (covariance and cholesky) seem close, but they don't match exactly.  In particular, they diverge slightly when peak lock is reached.  This could be an actual bug, or it could be a result of the increased precision in the cholesky decomposition method.  Unit tests for these functions are needed.