Work on code reorg
In this branch, I want to do a bit of a code reorg.  Specifically, I want to pull all of the "spacecraft-specific" stuff into seperate spacecraft classes, i.e. a Chandra class, a Suzaku class, and other classes if needed.  That way, the code can run somewhat agnostically as to what spacecraft it's dealing with.  Also, this will help a lot with code organization and neatness.